+++ 9ls1-fotoportal.php
@@ -2,13 +2,13 @@
 /**
  * Plugin Name: Aurora Fotoportal
  * Description: Aurora Fotoportal – kundeportal for fotoprosjekter, kontrakter, gallerier og levering.
- * Version: 0.7.1-dev.31-fix20
+ * Version: 0.7.1-dev.31-fix24
  * Author: 9Ls1 Digital
  * Text Domain: 9ls1-fotoportal
  */
 if (!defined('ABSPATH')) exit;
 
-define('NLS1_FOTOPORTAL_VERSION', '0.7.1-dev.31-fix20');
+define('NLS1_FOTOPORTAL_VERSION', '0.7.1-dev.31-fix24');
 define('NLS1_FOTOPORTAL_PLUGIN_DIR', plugin_dir_path(__FILE__));
 define('NLS1_FOTOPORTAL_PLUGIN_URL', plugin_dir_url(__FILE__));